requirements.txt
fastapi
uvicorn
ebooklib
beautifulsoup4
openai
python-docx
python-multipart
python-dotenv
